relocate task_output methods into pipeline_utils & prototype v2 of pipeline component
