+++ src/SOPRANO/pipeline_utils.py
@@ -1,6 +1,25 @@
 import pathlib
+from datetime import datetime
 
 from SOPRANO.objects import Parameters
+
+
+def time_output():
+    now = datetime.now()
+    return now.strftime("%d/%m/%Y %H:%M:%S")
+
+
+def task_output(msg):
+    print(f"[{time_output()}] {msg}")
+
+
+def is_empty(path: pathlib.Path) -> bool:
+    """
+    Checks whether file at path has size of zero
+    :param path: pathlib Path object
+    :return: True if path is empty else False
+    """
+    return path.stat().st_size == 0
 
 
 class MissingDataError(Exception):
@@ -9,6 +28,12 @@ class MissingDataError(Exception):
 
 class SOPRANOError(Exception):
     pass
+
+
+def _check_paths(*dependent_paths: pathlib.Path):
+    for path in dependent_paths:
+        if not path.exists():
+            raise MissingDataError(path)
 
 
 class _PipelineComponent:
@@ -32,10 +57,16 @@ class _PipelineComponent:
         pass
 
 
-def is_empty(path: pathlib.Path) -> bool:
-    """
-    Checks whether file at path has size of zero
-    :param path: pathlib Path object
-    :return: True if path is empty else False
-    """
-    return path.stat().st_size == 0
+class _PipelineComponent2:
+    msg = ""  # Will be printed to stdout with date/time stamp
+
+    def apply(self, params: Parameters):
+        task_output(self.msg)
+        self.check_ready(params)
+        self._apply(params)
+
+    def _apply(self, params: Parameters):
+        pass
+
+    def check_ready(self, *params: Parameters) -> None:
+        pass
